minor refactors to plugin tests and npm task output directory
