+++ apollo-gradle-plugin/src/test/groovy/com/apollostack/android/gradle/unit/ApolloIRGenTaskSpec.groovy
@@ -1,8 +1,8 @@
-package com.apollostack.android.gradle.project
+package com.apollostack.android.gradle.unit
 
 import com.apollostack.android.gradle.ApolloCodeGenInstallTask
 import com.apollostack.android.gradle.ApolloIRGenTask
-import com.apollostack.android.gradle.Utils
+import com.apollostack.android.gradle.ApolloPluginTestHelper
 import org.gradle.testfixtures.ProjectBuilder
 import spock.lang.Specification
 
@@ -10,10 +10,10 @@ class ApolloIRGenTaskSpec extends Specification {
   def "creates tasks for default project variants that depend on apolloCodgenInstall task"() {
     setup:
     def project = ProjectBuilder.builder().build()
-    Utils.setupDefaultAndroidProject(project)
+    ApolloPluginTestHelper.setupDefaultAndroidProject(project)
 
     when:
-    Utils.applyApolloPlugin(project)
+    ApolloPluginTestHelper.applyApolloPlugin(project)
 
     def debugTask = project.tasks.getByName(String.format(ApolloIRGenTask.NAME, "Debug"))
     def releaseTask = project.tasks.getByName(String.format(ApolloIRGenTask.NAME, "Release"))
@@ -26,10 +26,10 @@ class ApolloIRGenTaskSpec extends Specification {
   def "creates a top-level generateApolloIR task that depends on the variant tasks in a default project"() {
     setup:
     def project = ProjectBuilder.builder().build()
-    Utils.setupDefaultAndroidProject(project)
+    ApolloPluginTestHelper.setupDefaultAndroidProject(project)
 
     when:
-    Utils.applyApolloPlugin(project)
+    ApolloPluginTestHelper.applyApolloPlugin(project)
     def generateApolloIR = project.tasks.getByName(String.format(ApolloIRGenTask.NAME, ""))
 
     then:
