+++ apollo-sample/.gitignore
@@ -1,1 +1,2 @@
 /build
+package.json
